+++ lend-me-your-book/src/main/java/com/education/lendmeyourbook/repositories/StateRepository.java
@@ -3,6 +3,6 @@ package com.education.lendmeyourbook.repositories;
 import com.education.lendmeyourbook.entities.State;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StateRepository extends JpaRepository<State,Long> {
+public interface StateRepository extends JpaRepository<State,Integer> {
     public State findByName(String name);
 }
